+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Correção de _bug_ no dialeto VisuAlg quanto a ler e definir valores em matrizes: https://github.com/DesignLiquido/delegua/pull/575;
 - Correção de _bug_ em Delégua quanto a aglutinação de argumentos de funções: https://github.com/DesignLiquido/delegua/pull/570;
+- Novo analisador semântico: Mapler: https://github.com/DesignLiquido/delegua/pull/572;
 - Ajustes na gramática de Delégua para reconhecer tipos de estruturas de dados com acentos.
 
 ## 0.9.19